dynamic card props
