+++ containers/food/component/discount/Discount.tsx
@@ -1,14 +1,16 @@
-import { BigDiscountCard, BigDiscountCardData } from "./BigDiscountCard";
+// src/components/Discount/Discount.tsx
+import { BigDiscountCard } from "./BigDiscountCard";
+import { BigDiscountCardData, SmallDiscountCardWhiteData, SmallDiscountCardBlackData } from "@/constants/data";
 import "./discount.css";
-import { SmallDiscountCardBlack, SmallDiscountCardBlackData } from "./SmallDiscountCardBlack";
-import { SmallDiscountCardWhite, SmallDiscountCardWhiteData } from "./SmallDiscountCardWhite";
+import { SmallDiscountCardBlack } from "./SmallDiscountCardBlack";
+import { SmallDiscountCardWhite } from "./SmallDiscountCardWhite";
 
 const Discount = () => {
   return (
     <section className="discount_container">
       <div className="discount-food_frame">
         <div className="big_discount-card">
-          <BigDiscountCard data={BigDiscountCardData} />
+          <BigDiscountCard data={BigDiscountCardData} /> {/* Pass data as props */}
         </div>
         <div className="small_discount-card">
           <SmallDiscountCardWhite data={SmallDiscountCardWhiteData} />
@@ -18,4 +20,5 @@ const Discount = () => {
     </section>
   );
 };
+
 export default Discount;
